# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Rootcause API",
    "version": "1.0.0",
    "description": "API for Rootcause dashboard, transaction monitoring, and investigations. This detailed specification can be used for integration with Model Context Protocol (MCP) or standard OpenAPI clients."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local server"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "description": "Returns the health status of the API. Useful for ensuring the server is up and running.",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "API is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "status": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/investigations": {
      "post": {
        "summary": "Start Investigation",
        "description": "Initiates a root cause investigation for a specific failed transaction ID.",
        "operationId": "startInvestigation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["transactionId"],
                "properties": {
                  "transactionId": {
                    "type": "string",
                    "description": "The unique identifier of the transaction to investigate"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Investigation request successfully received",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "transactionId": { "type": "string" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed, likely missing or invalid transactionId",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/transactions/list": {
      "get": {
        "summary": "List Transactions",
        "description": "Retrieves all monitored transactions, including both successful and failed payments.",
        "operationId": "listTransactions",
        "responses": {
          "200": {
            "description": "A comprehensive list of transactions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Transaction"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/transactions/{id}": {
      "get": {
        "summary": "Get Transaction by ID",
        "description": "Retrieves a specific transaction by its unique identifier.",
        "operationId": "getTransactionById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the transaction to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "$ref": "#/components/schemas/Transaction"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Transaction not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/transactions/{id}/investigate": {
      "post": {
        "summary": "Investigate Transaction via AI",
        "description": "Initiates an AI-driven root cause investigation for a specific transaction ID. Returns a Server-Sent Events (SSE) stream.",
        "operationId": "investigateTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the transaction to investigate",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE Stream of AI investigation chunks",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "Event stream data payload prefixed with 'data: '"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/transactions/incident": {
      "get": {
        "summary": "List Incidents",
        "description": "Retrieves only the failed transactions, which are classified as incidents requiring attention.",
        "operationId": "listIncidents",
        "responses": {
          "200": {
            "description": "A list of failed transactions (incidents)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Transaction"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/dashboard/stats": {
      "get": {
        "summary": "Get Dashboard Stats",
        "description": "Retrieves high-level overview statistics for the main dashboard, including total transactions, failure rates, and critical incidents.",
        "operationId": "getDashboardStats",
        "responses": {
          "200": {
            "description": "Dashboard statistics data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "totalTransactions": { "type": "number", "description": "Total volume of transactions monitored" },
                        "totalTransactionsChangePct": { "type": "number", "description": "Percentage change in transaction volume" },
                        "failedTransactions": { "type": "number", "description": "Total number of failed transactions" },
                        "failedTransactionsPct": { "type": "number", "description": "Percentage of transactions that failed" },
                        "activeIncidents": { "type": "number", "description": "Currently active incident count" },
                        "activeIncidentsCreatedToday": { "type": "number", "description": "Incidents created in the last 24h" },
                        "criticalIncidents": { "type": "number", "description": "Number of critical priority incidents" },
                        "criticalRequiringAttention": { "type": "number", "description": "Critical incidents pending resolution" }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/dashboard/severity": {
      "get": {
        "summary": "Get Severity Breakdown",
        "description": "Retrieves an aggregate count of incidents categorized by severity (CRITICAL, HIGH, MEDIUM, LOW).",
        "operationId": "getSeverityBreakdown",
        "responses": {
          "200": {
            "description": "Array containing severity breakdowns",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "severity": { "type": "string", "enum": ["CRITICAL", "HIGH", "MEDIUM", "LOW"] },
                          "count": { "type": "number" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/dashboard/activity": {
      "get": {
        "summary": "Get Activity Chart Data",
        "description": "Retrieves time-series activity metrics (successful vs failed vs suspicious) over a specified time range.",
        "operationId": "getActivity",
        "parameters": [
          {
            "name": "range",
            "in": "query",
            "description": "Time range for the activity graph (e.g., 24h, 7d, 30d). Defaults to 24h if not specified.",
            "required": false,
            "schema": {
              "type": "string",
              "default": "24h"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Time-series activity data suitable for charting",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "label": { "type": "string", "description": "Formatted time label (e.g. '10 AM', 'Mon')" },
                          "timestamp": { "type": "string", "format": "date-time", "description": "ISO timestamp of the data point" },
                          "successful": { "type": "number", "description": "Count of successful transactions" },
                          "failed": { "type": "number", "description": "Count of failed transactions" },
                          "suspicious": { "type": "number", "description": "Count of suspicious transactions" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Transaction": {
        "type": "object",
        "description": "Represents a payment transaction object.",
        "properties": {
          "id": { "type": "string", "description": "Unique transaction identifier (e.g. pay_abc123)" },
          "userId": { "type": "string", "description": "Identifier of the user who initiated the transaction" },
          "amount": { "type": "number", "description": "Transaction amount" },
          "currency": { "type": "string", "description": "Three-letter currency code (e.g. USD, INR)" },
          "status": {
            "type": "string",
            "enum": ["SUCCESS", "FAILED", "PENDING"],
            "description": "Current status of the transaction"
          },
          "timestamp": { "type": "string", "format": "date-time", "description": "ISO 8601 timestamp of when the transaction occurred" },
          "paymentMethod": { "type": "string", "description": "Method of payment (e.g. CREDIT_CARD, UPI)" },
          "gateway": { "type": "string", "description": "Payment gateway used" },
          "errorMessage": { "type": "string", "description": "Error message details if the transaction failed" },
          "errorCategory": { "type": "string", "description": "Categorized reason for failure (e.g. FRAUD_SUSPECTED)" }
        },
        "required": ["id", "userId", "amount", "currency", "status", "timestamp", "paymentMethod"]
      }
    }
  }
}